move to delphi/eval
